+++ django_perf_rec/__init__.py
@@ -19,4 +19,4 @@ if pytest is not None:
 
 from .api import TestCaseMixin, get_record_name, get_perf_path, record  # noqa: F401
 
-__version__ = '2.2.0'
+__version__ = '3.0.0'