+++ HISTORY.rst
@@ -8,6 +8,9 @@ Pending release
 
 .. Insert new release notes below this line
 
+3.0.0 (2018-07-17)
+------------------
+
 * Don't replace columns in ORDER BY, GROUP BY and HAVING clauses.
 
 2.2.0 (2018-01-24)